requirements.txt
Flask
Flask-Login
oauthlib
requests
mysql-connector-python
pypdf
reportlab
squareup
azure-storage-file-share
